requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmdstanpy==1.0.8
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
convertdate==2.4.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
discord.py==2.1.0
entrypoints==0.4
ephem==4.1.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.3.3
greenlet==2.0.1
hijri-converter==2.2.4
holidays==0.18
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
load-dotenv==0.1.0
LunarCalendar==0.0.9
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.4
nbformat==5.7.1
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.11.0
plotly-express==0.4.1
prompt-toolkit==3.0.36
prophet==1.1.1
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pycron==3.0.0
Pygments==2.14.0
PyMeeus==0.5.12
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pywin32==305
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
setuptools-git==1.2
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
statsmodels==0.13.5
table2ascii==1.1.0
tenacity==8.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
urllib3==1.26.13
wcwidth==0.2.5
yarl==1.8.2
